+++ .gitignore
@@ -86,3 +86,4 @@ src/AssemblySharedInfoGenerator/AssemblySharedInfo.cs
 /src/Libraries/PythonNodeModels/*.resources
 /test/core/dsevaluation/DSFiles/test.ds
 GregStandaloneAuth.dll
+/src/Libraries/Watch3DNodeModels/*.resources
