--- src/Libraries/DynamoWatch3D/Properties/AssemblyInfo.cs
@@ -1,13 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("DynamoWatch3D")]
-[assembly: AssemblyCulture("")]
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("b48c0f05-8716-44d6-9c5e-20ed731a6f0d")]
-[assembly: InternalsVisibleTo("DynamoCoreTests")]
-[assembly: InternalsVisibleTo("DynamoCoreWpfTests")]
