+++ src/Libraries/Watch3DNodeModelsWpf/packages.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Cyotek.Drawing.BitmapFont" version="1.0.2.0" targetFramework="net45" />
+  <package id="HelixToolkit" version="2015.1.629" targetFramework="net45" />
+  <package id="HelixToolkit.Wpf" version="2015.1.629" targetFramework="net45" />
+  <package id="HelixToolkit.Wpf.SharpDX" version="2015.1.629" targetFramework="net45" />
+  <package id="SharpDX" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.D3DCompiler" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Direct2D1" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Direct3D11" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Direct3D11.Effects" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Direct3D9" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.DXGI" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Toolkit" version="2.6.3" targetFramework="net45" />
+  <package id="SharpDX.Toolkit.Graphics" version="2.6.3" targetFramework="net45" />
+</packages>
